Remove the deprecated `otel/metric/global` pkg
