+++ internal/global/instruments.go
@@ -12,14 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package global // import "go.opentelemetry.io/otel/metric/internal/global"
+package global // import "go.opentelemetry.io/otel/internal/global"
 
 import (
 	"context"
 	"sync/atomic"
 
 	"go.opentelemetry.io/otel/attribute"
-	oGlob "go.opentelemetry.io/otel/internal/global"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/instrument"
 )
@@ -44,7 +43,7 @@ var _ instrument.Float64ObservableCounter = (*afCounter)(nil)
 func (i *afCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableCounter(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -72,7 +71,7 @@ var _ instrument.Float64ObservableUpDownCounter = (*afUpDownCounter)(nil)
 func (i *afUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableUpDownCounter(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -100,7 +99,7 @@ var _ instrument.Float64ObservableGauge = (*afGauge)(nil)
 func (i *afGauge) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64ObservableGauge(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -128,7 +127,7 @@ var _ instrument.Int64ObservableCounter = (*aiCounter)(nil)
 func (i *aiCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableCounter(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -156,7 +155,7 @@ var _ instrument.Int64ObservableUpDownCounter = (*aiUpDownCounter)(nil)
 func (i *aiUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableUpDownCounter(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -184,7 +183,7 @@ var _ instrument.Int64ObservableGauge = (*aiGauge)(nil)
 func (i *aiGauge) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64ObservableGauge(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -212,7 +211,7 @@ var _ instrument.Float64Counter = (*sfCounter)(nil)
 func (i *sfCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64Counter(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -238,7 +237,7 @@ var _ instrument.Float64UpDownCounter = (*sfUpDownCounter)(nil)
 func (i *sfUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64UpDownCounter(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -264,7 +263,7 @@ var _ instrument.Float64Histogram = (*sfHistogram)(nil)
 func (i *sfHistogram) setDelegate(m metric.Meter) {
 	ctr, err := m.Float64Histogram(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -290,7 +289,7 @@ var _ instrument.Int64Counter = (*siCounter)(nil)
 func (i *siCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64Counter(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -316,7 +315,7 @@ var _ instrument.Int64UpDownCounter = (*siUpDownCounter)(nil)
 func (i *siUpDownCounter) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64UpDownCounter(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
@@ -342,7 +341,7 @@ var _ instrument.Int64Histogram = (*siHistogram)(nil)
 func (i *siHistogram) setDelegate(m metric.Meter) {
 	ctr, err := m.Int64Histogram(i.name, i.opts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 		return
 	}
 	i.delegate.Store(ctr)
