+++ internal/global/meter.go
@@ -12,14 +12,13 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package global // import "go.opentelemetry.io/otel/metric/internal/global"
+package global // import "go.opentelemetry.io/otel/internal/global"
 
 import (
 	"container/list"
 	"sync"
 	"sync/atomic"
 
-	oGlob "go.opentelemetry.io/otel/internal/global"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/instrument"
 )
@@ -33,11 +32,6 @@ type meterProvider struct {
 	meters map[il]*meter
 
 	delegate metric.MeterProvider
-}
-
-type il struct {
-	name    string
-	version string
 }
 
 // setDelegate configures p to delegate all MeterProvider functionality to
@@ -334,7 +328,7 @@ func (c *registration) setDelegate(m metric.Meter) {
 
 	reg, err := m.RegisterCallback(c.function, insts...)
 	if err != nil {
-		oGlob.GetErrorHandler().Handle(err)
+		GetErrorHandler().Handle(err)
 	}
 
 	c.unreg = reg.Unregister
